requirements.txt
groq==1.0.0
python-dotenv==1.2.1
